# app/agency-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Agency service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/agency-service",
    "paths": {
        "/v1/agencies": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agencies"
                ],
                "summary": "Get agencies",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Take",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "Query",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agencies response",
                        "schema": {
                            "$ref": "#/definitions/dto.AgenciesResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agencies"
                ],
                "summary": "Create an agency",
                "parameters": [
                    {
                        "description": "Agency data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAgencyDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Agency response",
                        "schema": {
                            "$ref": "#/definitions/dto.AgencyDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/agencies/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agencies"
                ],
                "summary": "Update an agency",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Agency ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Agency data",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAgencyDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agency response",
                        "schema": {
                            "$ref": "#/definitions/dto.AgencyDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/artists": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Get artists",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Take",
                        "name": "take",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "Query",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Agency ID",
                        "name": "agencyId",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Is User Followed",
                        "name": "isUserFollowed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artists response",
                        "schema": {
                            "$ref": "#/definitions/dto.ArtistsResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Create an artist",
                "parameters": [
                    {
                        "description": "Create artist request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateArtistDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist response",
                        "schema": {
                            "$ref": "#/definitions/dto.ArtistDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/artists/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Update an artist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update artist request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateArtistDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist response",
                        "schema": {
                            "$ref": "#/definitions/dto.ArtistDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/artists/{id}/subscription": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Artists"
                ],
                "summary": "Handle a subscription to Artist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription response",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscriptionResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AgenciesResponseDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AgencyDTO"
                    }
                },
                "skip": {
                    "type": "integer"
                },
                "take": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.AgencyDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string",
                    "example": "https://image-service.com/image.png"
                },
                "name": {
                    "type": "string",
                    "example": "agency name"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "dto.ArtistDTO": {
            "type": "object",
            "properties": {
                "agencyId": {
                    "type": "integer",
                    "example": 1
                },
                "debutDate": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string",
                    "example": "https://image-service.com/image.png"
                },
                "name": {
                    "type": "string",
                    "example": "Artist John Doe"
                },
                "numFollowers": {
                    "type": "integer",
                    "example": 100
                },
                "status": {
                    "type": "string",
                    "example": "disband"
                }
            }
        },
        "dto.ArtistsResponseDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ArtistDTO"
                    }
                },
                "skip": {
                    "type": "integer"
                },
                "take": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateAgencyDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "imageUrl": {
                    "type": "string",
                    "example": "https://image-service.com/image.png"
                },
                "name": {
                    "type": "string",
                    "example": "agency name"
                }
            }
        },
        "dto.CreateArtistDTO": {
            "type": "object",
            "required": [
                "agencyId",
                "debutDate",
                "name"
            ],
            "properties": {
                "agencyId": {
                    "type": "integer",
                    "example": 1
                },
                "debutDate": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "imageUrl": {
                    "type": "string",
                    "example": "https://image-service.com/image.png"
                },
                "name": {
                    "type": "string",
                    "example": "Artist John Doe"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Some error message"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "dto.SubscriptionResponseDTO": {
            "type": "object",
            "properties": {
                "artistId": {
                    "type": "integer",
                    "example": 10
                },
                "isFollowed": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.UpdateAgencyDTO": {
            "type": "object",
            "properties": {
                "imageUrl": {
                    "type": "string",
                    "example": "https://image-service.com/image.png"
                },
                "name": {
                    "type": "string",
                    "example": "agency name"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "dto.UpdateArtistDTO": {
            "type": "object",
            "properties": {
                "agencyId": {
                    "type": "integer",
                    "example": 1
                },
                "debutDate": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "imageUrl": {
                    "type": "string",
                    "example": "https://image-service.com/image.png"
                },
                "name": {
                    "type": "string",
                    "example": "Artist John Doe"
                },
                "status": {
                    "type": "string",
                    "example": "disband"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}